data_science/requirements.txt
opencv-python>=4.5.0
azure-storage-blob>=12.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0 